Fixed Mouse Left double-clicking misbehaving on Windows
VLC handles double-click differently on Linux and Windows.

On Linux, you always end up in the same state (either playing or paused) after
double clicking that you were in before the double clicking. It goes like this:

first click -> play/pause -> second click -> pause/play (note the reverse order)

On Windows, however, you end up in the opposite state after double clicking
that you were in before the double clicking. It goes like this:

first click -> play/pause -> second click -> nothing changes

This behaviour is annoying, because if you want to full-screen with the double-
click shortcut, then you have to do a third click to return video to the same
state it was before.

This commit makes Windows double-clicking behave like it does on Linux.

Note that this seems to apply only to Mouse Left, it's likely that double-
clicking behaviour of other buttons will still be different between Linux and
Windows. For example, this is the case with Mouse Middle button. Windows VLC
simply doesn't register the second click of it until the double-click time
window passes.