Removed support for VLC 3.0.0
There is no VLC 3.0.0 to begin with, at least yet. What is called VLC
3.0.0 is development branch of VLC repository, which keeps constantly
changing. Because of it changing, the plugins built for it at one point
in time no longer work anymore. I'm also not able to build VLC 3.0.0
because the tarballs with prebuilt dependencies provided by the VLC
team are provided only for released versions, which 3.0.0 is not.