Made README more Windows user friendly
For some reason Windows users think that the "n bit" in the table
refers to the bitness of their Windows version, not to the bitness of
VLC they are using. Since majority of Windows installs are 64-bit and
the VLC download page serves 32-bit VLC by default, this leads to a lot
of "why does this plugin not work?" questions, some of which even made
it to superuser dot com... Well, hopefully this change makes it clear.