Merged source code for the two plugin versions into one file
It was getting a bit of a pain to maintain two almost identical source files
with slight differences, so I just merged them together in a way that would work
both for VLC 2.1 and VLC 2.2.
This change requires VLC 2.1 to additionally check the plugin's checkbox in
Advanced Preferences -> Video -> Filters -> Pause/Play video on mouse click,
just like in VLC 2.2, making the same usage instructions work for both plugins.