Improve handling of settings
We were using the wrong type for mouse button settings -- we used
string_list when we actually wanted to use integer_list. Instead of
having to encode the integer value of the selected mouse button as a
printable character we can now store the integer value directly. We no
longer store the mouse button values VLC uses internally, we now store
an index into an array of mouse buttons. This gives us some flexibility,
such as an ability to re-arrange the button list, and it might come in
handy if VLC chooses to change values for the mouse button constants.

This is a breaking change, in the way that the previously stored plugin
settings are not compatible with this change. User's previous mouse
selection won't load properly and will be reset to its defaults. Users
will have to re-configure the plugin if they don't like the defaults.

Also, since we are breaking the config compatibility anyway, some of
settings were renamed to more appropriate names, which is also a
breaking change.