Fix the option for ignoring double clicks
That option had a rather weird behavior. It was made so that the video
wouldn't pause on the double click, when a used supposedly wants to
fullscreen the video instead. It would ignore the first click, assuming
that it could be the first click of a double click, and it would wait
the specified <double click interval> milliseconds for the second click.
If there was no second click when the time runs out, the plugin realizes
that it was a single click and pauses/plays the video. However, if a
second click did happen within that time interval, then it won't
pause/play it.

It all makes sense aside from the user having to specify the double
click interval. It's actually system dependent and VLC provides no
facility to query what that interval is, so we rely on the user to
provide a good guess. This can result in a weird behavior if user's
guess is not good enough, as the double-click-to-fullscreen happens
using system's double click interval. For example, if the system double
click interval is 300ms, but the user set it in the plugin as 200ms,
then when the user double clicks with the clicks being 250ms apart, the
plugin won't recognize it as a double click causing the video to
pause/play, but the system will recognize it as a double click, causing
the video to fullscreen at the same time.

With this change we eliminate the OS-specific double click interval and
instead make fullscreen happen at the user-specified double click
interval, which solves this interval inconsistency issue. It
additionally allows the user to control the double-click-to-fullscreen
interval, making it shorter or longer, which wasn't possible before.