Switch from Travis-CI to GitHub Actions
Currently lacks nightly VLC 4.0 release publishing support which
Travis-CI had.

The build times could improve greatly with some caching:
  - On Windows: cache the Docker image we build and re-use it for all
  Windows jobs
  - On Linux: cache per-job Docker images and VLC downloads
  - On macOS: cache VLC 2.1, 2.2 and 3.0 downloads

However, it's a bit of work to implement all this caching and it would
add complexity to the workflow file and build scripts, so given that
it's just me working in this repo, that my commit frequency is low (the
plugin is pretty much feature complete at this point, there is no real
development happening, just maitenance) and that the entire build
completes in reasonable 4-5 minutes, there is little reason to implement
caching.